Added example of async/await downside
I added an example that shows the downsides of using async/await. There is a function that takes milliseconds as a parameter, and sets a timeout. I used this function in the initialize(), and use the timeout to set a console.log. I also included a fetch request. If we used async/await, we would have to wait 2 seconds and then the website would make the fetch request. This example shows that we need to stick to .then statements and use promises instead of async/await.