Using async/await with CRUD (POST)
I added an example of using async/await while fetching API, and doing a CRUD operation. In this example I posted a grade (static) which was inputted by the user in the saveGrade function.

I then used DOM selection to access the HTML elements which the user interacted with. I accessed the grades-form element which allows me to add an event listener on submit. I also accessed the user-grade input element that allows me to access the value the user input.

I then added an event listener for a "submit" on the form element, that whenever a user submits it will first preventDefault() which makes sure the website doesn't refresh, and then it runs the saveGrade function with the userGrade value that was inputed.