Handling rejected awaits with Try...Catch blocks
I added 2 different ways to handle rejected awaits with Try...Catch blocks. The first way wraps the getUnreadCount function in another async function named init. This init function implements a try catch block which tries the getUnreadCount function by adding await and storing it to a constant. If there is an error then it catches the error and it console.error. This way allows for there to be more control in handling the errors. It will do it case by case.

The second way wraps the getUnreadCount function directly with a try catch block. This simply means that it will handle all errors within that function the same way. To do this we have to throw an error.