Example of a FetchWrapper Class, used with async/await while fetching API and handling errors
In this code addition, I created a class named FetchWrapper, where I am able to handle CRUD operations. In this example, I only added a get operation. I incorporated this class with async/await instead of using .then.

I then created a new use of this FetchWrapper to get notifications from the API, while handling errors. I did this in the checkForNotifications function using async/await. In the try block I await for a response from the get operation. I then assign it to a const and pass the data I need to a different function with handles the response. In my catch statement I simply catch any error, and also call another function that handles my error as well.

The function which handles the output message to the user is function showNewNotifications. This function checks if there is no data (which means that there was an error caught), if so it will console.log to the user that the notifications could not be displayed. Else, it will log the user the count(data) in a display message.