Updated and finished the FetchWrapper Class
I updated my FetchWrapper class which contains all of my API operations (CRUD). I added put post delete using async and await in them instead of working directly with promises using the .then syntax. I added body as a parameter since all of the last 3 operations are changing the content of the API someway. I show that these operations all include the method which includes what operation is being done, the headers which informs that the content type is a json file, the body then is being transformed into JSON format, and finally we return the response into json format.