Added note on Dynamic Imports
I am diving more on using different libraries and the power of importing and exportin. In this example I show Dynamic imports on user Interaction. I do this by working with the DOM on a button element and adding an event listener for a click. I also incorporated async/await into this.

In the second example I noticed that since I was receiving an object from the import, I can work with destructuring. I show that I destructure the init function from the module object.