I added regular Dynamic imports
I am showing an example of Dynamic imports that would then be used inside of a function. I am importing the lodash library which has a default export that is debounce. I am initializing a const and setting that default export equal to it, which I will then use in the same function.

I am doing the same with ./helpers.js, which stores the default export "Helpers" in a Helpers const, and then accesses the getDate named export from module and sets the value of that in a getDate const.