Example on Closures
I added an example on the Closures topic. I want to display confetti to a user once the user clicks on a button. If the confetti is already shown I want to make sure that it doesn't show it again.

I use the concept of Closures here, by accessing the isConfettiShown variable inside of my callback function, which checks the value of it in an if condition.

I then CHANGE the isConfettiShown value to a different value.

I am able to do this because isConfettiShown is defined in the global scope.