Another example of using Closures
I added another example where I access a variable in the global scope. I want to run the startGame() only one when the user presses on the button, but I want the keep calling the logAnalytics() everytime.

I change access the isCalled value in an if condition and then I change the isCalled value.