I added an example that uses sets and loops
In this example I have a function which obtains all the possible outcomes of two dice being rolled at the same time, and storing those outcome sin a set. I use nested for loops to obtain the possible combinations and then I use set.add(firstRoll + secondRoll) which adds the combination of both rolls. This allows for the set to be unique and for there to be no duplication in numbers.