Added example of functional programming
In this example, I added an array of objects named users. I then use functional programming to filter out the users that are adults and are active.

I use chaining and link both filters together and set that equal to a constant. This will hold the new array of objects that will be created.

I take it to another level and refactor my code by making the arrow functions equal to a constant. I then use this constant inside the filter method.