Another example of functional programming
In this example, I have an array of objects which stores information about specific products. I want a new array of objects that contains the items that are above 100 in price, are low in stock, and I want their names.

To do this I used the filter method. I set my arrow functions equal to specific constants, and passed those constants are my arguments in the filter().