Another currying example
In this example I imported an array of objects named Users. I wanted to filter through that array and give me the products which where low in stock. I utilized currying in the equation filterByLowStock by first taking the lowStock parameter, then returning another function which takes products as a parameter, which then filters the array products into another array based on the condition I gave it.